# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MusicBox API",
        "description": "Documentation interactive de l'API MusicBox (Artistes, Albums, Chansons)",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/artists": {
            "get": {
                "tags": [
                    "Artists"
                ],
                "summary": "Liste des artistes",
                "operationId": "2c8aed1387cca1c65d98da7ee11ac926",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page pour pagination",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "genre",
                        "in": "query",
                        "description": "Filtrer par genre musical",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste paginée d'artistes"
                    }
                }
            },
            "post": {
                "tags": [
                    "Artists"
                ],
                "summary": "Créer un nouvel artiste",
                "description": "Ajoute un artiste dans la base de données",
                "operationId": "83d2888bdc6fac66b9a7cc81c4c950fb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "genre"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Freddie Mercury"
                                    },
                                    "genre": {
                                        "type": "string",
                                        "example": "Rock"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Artiste créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Freddie Mercury"
                                        },
                                        "genre": {
                                            "type": "string",
                                            "example": "Rock"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requête invalide"
                    }
                }
            }
        },
        "/api/artists/{id}": {
            "put": {
                "tags": [
                    "Artists"
                ],
                "summary": "Mettre à jour un artiste",
                "description": "Met à jour les informations d’un artiste existant",
                "operationId": "0f6af6d6ced6c49d7125ca25a311215c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'artiste",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Mick Jagger"
                                    },
                                    "genre": {
                                        "type": "string",
                                        "example": "Rock"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Artiste mis à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Mick Jagger"
                                        },
                                        "genre": {
                                            "type": "string",
                                            "example": "Rock"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Artiste non trouvé"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Artists"
                ],
                "summary": "Supprimer un artiste",
                "description": "Supprime un artiste de la base de données",
                "operationId": "4d8899f5b7d1243c5e3eaf300f05f9f7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'artiste",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Artiste supprimé avec succès"
                    },
                    "404": {
                        "description": "Artiste non trouvé"
                    }
                }
            }
        },
...
    },
...
}